--- Material.Dialog/Commands/RelayCommand.cs
@@ -1,34 +1,0 @@
-﻿using System; 
-using System.Windows.Input;
-
-namespace Material.Dialog.Commands
-{
-    public class RelayCommand : ICommand
-    {
-        private Action<object> execute;
-        private Func<object, bool> canExecute;
-        public event EventHandler CanExecuteChanged;
-
-        public RelayCommand(Action<object> execute, Func<object, bool> canExecute = null)
-        {
-            this.execute = execute;
-            this.canExecute = canExecute;
-        }
-
-        public bool CanExecute(object parameter)
-        {
-            return this.canExecute == null || this.canExecute(parameter);
-        }
-
-        public void Execute(object parameter)
-        {
-            this.execute(parameter);
-        }
-
-        public void RaiseCanExecute()
-        {
-            var handler = CanExecuteChanged;
-            handler?.Invoke(this, new EventArgs());
-        } 
-    }
-}
